kboot: Prevent NULL pointer dereference in dt_set_display()
dt_set_display() relied on display_get_config() returning the default
config entry as it is only interested on the device tree alias for the
primary dcp. This is always "dcp" except for devices with a config.
Since commit 8508526 ("display: contain all display init logic within
display.c") it however returns NULL if ADT dcp node specified in the
config does not exists. This is the case for M2 Pro and Max macbooks
which use "dcp0" instead of "dcp".

Fixes: 8508526 ("display: contain all display init logic within display.c")
Signed-off-by: Janne Grunau <j@jannau.net>